Exercise 4.4 multi-argument functions
Significantly more straightforward than expected.

Though, mistakenly using Expr rather than AtExpr for the base case in AtExprs led to some confusion.
Specifically, with an Expr base caes something like "n - 1" is interpreted as n(-1), as -1 is a well formed expression and can be pused to the list.
The general structure of the language prevents this issue from spreading by requiring deliniation of the start of an expression.
A small test has been added for this observation.