Exercise 4.5
Fairly direct.

That strings (rather than chars only) are permitted in tokens was a guess based on the option to use regex.
This seems to work fine.

I'm surprised negation doesn't follow as the next exercise, though it's easy to implement as a function.
I.e. negation x = if x = 0 then 1 else 0, etc.