Exercise 4.6, tuple constructors
At least for now…

Tests pass, and things seem roughly in order.

Still, the implementation is gnarly, an in particular the checks on whether or not to evaluate a value.
Though, in part this may be to me not fully understanding F# formatting rules, and as a result failing to clarify some of the control flow.

Details aside, an interesting exercise.
Highlighting consequences of earlier decisions and allowing for some implementation consideration.