Exercise 4.7, typed parameters
Fairly straightforward.

The suggestion to delete the `eval` function was ignored, given the use of the function in tests.
So, it has also been updated to work with multiple parameters.
The changes were quite light, given most of the conceptual work was done when modifying the untyped language.

In line with the above, tests have been updated.

The underlying implementation of List.methods is unknown to me.
So, perhaps forall, zip, etc. could be optimised.
Still…