Exercise 5.3, merging lists
Mostly simple.

Mostly, as type inference for functional parameters in C++ isn't so clear to me.

I expect these exercises are primarily designed as introductions to function passing.
And, in this respect, these are lessons already learnt --- though likely for other reads too given the amount of F# inspected so far.