Exercise 5.7, typed lists
As there's no guidance on how to adapt evaluation to lists, changes have only been made to the checker.

Implementation is, I expect, quite standard.

List construction requires specifying a single type for all elements in the list to conform to.
Checking, then, quite literally, ensures all elements in the list conform to the type.