Exercise 4.9 / 5.7 (and a little of 4.8), lists
Exercises 4.8 and 4.9 were deferred to help reinforce the ideas.

It seems 4.9 and 5.7 are the same.
And, with the context of 4.8 the respective list structure is clearer.

The revised code recursively checks the list.
Specifically, the idea is to determine the head type, and then match against the tail.
If nil, then return the head type, if a list and recursive application returns the same type, then that type and otherwise a mismatch, and otherwise the list is malformed.

This seems to be what is intended, or at least is much closer, and so the previous approach to lists has been removed.