Exercise 4.7, match expressions
Things may have left the tracks… though I'm unsure

The idea with match statements is to bind head and tail in the environment, so the match expression has access.
This means the environment should support lists, that variables should bind lists, and so on.

My preferred method to implement the above is to generalise the return type of eval from int to value.
To this a list can be added, and in the process booleans are also added.

This cascades a little into binary ops, as now it makes little sense to coerce boolean ops to ints, and to wrapping the final return type.
Though, once implemented the changes are fairly minimal and have the nice side effect of almost typed returns.

To test a rough instance of exists is implemented, though which only checks if 1 exists in a list of ints.