Exefcise 7.1, MiniC interpretation
The AST specification is updated with a prefix to avoid namespace conflicts.
And, interpretation / evaluation / etc. has been updated to take an addition `out` parameter.
The `out` parameter is a string ref, and print statements have been reinterpreted to append to the value of the parameter.
With this, tests are relatively straightforward, though string matching.