Exercise 7.5, lex and parse pre inc/dec
Only minor modifications are needed.

Inc and dec are set with higher precedence than plus and minus, as this should clearly resolve cases such as `a -- b`, should be interpreted as `a (--b)`, or at least seems to make most sense to do.

And, in order to prevent ``--1`` the `Access` parse is used.