Exercise 7.6, compound assignments
A somewhat elegant approach would be to define construction without introducing an existing token.
So, e.g., `+=` would be parsed as PLUS ASSIGN and things follow from here.
This, though, raises the issue of `+ =`.
The separation isn't accepted in C, and it makes sense to prevent acceptance here.
Still, as the lexer is whitespace indifference, `+=` and `+ =` will be parsed the same.
So, it seems fresh tokens are the way to go.
Well, it's almost interesting…