Exercise 8.3, inc dec code
The exercise trivialised this a little, though still entertaining.

Though, some things were learnt.

Specifically, the interpretation of ++arr[++i].

Here, `++` requires an lvalue.
So much is fine.
Though, does this require the array is an array of lvalues?
No, it does not, as arr[++i] is itself the lvalue arr + i.
There's no evaluation, and so the increment goes through fine.
It's neat in a statement oriented language, though I can see why it's avoided in Rust, as it's a compound with brackets binding tighter --- which suggests evaluation may occur.