Exercise 8.5, compilation of ? :
Quite a straightforward approach.

Syntax is handled with a new type `Ite` and adding tokens for `?` and `:`.
The only part to note is precedence.
Both `?` and `:` are set with low precedence to avoid breaking apart otherwise fine expressions.
Though, not much thought was given, and perhaps some additional tests would show the level should be lowered (or raised).

Interpretation of Ite uses F#'s if / then / else, so no notes.

Compilation is, also, simple.
Compile the condition, continue to the y case if the result is non-zero and then jump over the n case.
Otherwise, jump to the n case and then end.