Extend access and assign to compilation
And with some simplification.

Previously, access and assign was handled directly through an AccessAssign type.

This is not at all required, as, e.g., `a[i] += 1` can be seen as shorthand for `a[i] = a[i] + 1`.
So, things have been updated so that the parser performs this rewrite and everything else is simplified where possible.

Of course, this isn't the most efficient implementation.
As, here with the rewrite, there are two reads of `a[i]` which could be simplified to a duplication.
Still, the generality given is a worthwhile tradeoff, I think, in this case.