Exercise 8.9, initialised declarations --- basics
Basic here means:

- Local only
- Interpretation only
- Restrictions not fully thought through.

For parsing, VardecA is introduced to cover initialised declarations.
Consider:

- int x = 1;
- int* x = &y;
- int* x = 1;
- int a[i] = y;

The first two are fine, the other two are not.
Or, there's good argument against the other two.
Specifically, the third initialised a pointer to an arbitrary address while the fourth initialises an extant location.

VardecA is quite restrictive.
Initialising a name to any expression is fine.
And, the only other permissible iniialisation is when a `*` is paired with a `&`.
The latter case could (and should) be generalised to any depth of pairing, though for simplicity it's kept to a depth of one.

As the parser distinguishes local and global variables, it's also simple to restrict things.

Of minor issue is whether the restrictions enforced when parsing should be enforced by the type system also.
This is not required at the moent, and might not be worthwhile.

That is, `int* x = 1` could be allowed.

The irony of syntactically gatekeeping semantically well formed constructs isn't lost.