Exercise 8.9, initialised declarations, refined a little
Rather than parsing names as strings, it makes a little more sense to parse names as access variables.
This way the intent is a little clearer, and failure cases for non-variable declaration initialisations are clearer.
As, with this implemented both the interpreter and compiler must match the access type to obtain a variable for declaration.

Compilation is very straightforward, though primarily as a result of the abstractions given.