Exercise 8.9, initialised declarations
Previous commits covered the local case.

For global initialisation only compilation is supported (following the task).

The idea is to parse global initialisation declarations as in the local case.
Then, at compilation time split the initialisation and declaration.
Initialisation happens as before, with all initialisations being placed at the bottom of the stack.
Declaration is deferred until the program is ran, and happens immediately.

Very roughly, a global of the form

int i = 3;

Is translated to:

int i;

int main() {
  i = 3;
}

Though in practice main is not called.

This is difficult to replicate in the interpreter.
Further, it's also difficult to perform immediate evaluation of global variables.
So, the interpreter isn't supported.

This seemed to be the best approach.

Performing immediate evaluation seemed difficult, esp. wrt. to fixing locations.

And, though it should be the case that the value of every assignment can be inferred at comp time, using a modified interpreter to evaluate and then storing the value seemed brittle.
In that, ideally there should be a single source of evaluation, etc.

---

In addition, the DecA type has been reverted to using strings.
This helps limit changes to existing functions.