listmachine revisions
Somewhat significant revisions to listmachine.

The primary goal was to remove dependency on 32 bit ints.
Independnece is acheived, primarily, by use of `intptr_t` to fix various datatypes to the size of a pointer.
The check on size now requires /at least/ 32bit pointers.

In addition, alias have been introduced for the type of a word and the type of an instruction.
These are used for clarity.

Similarly, most variables are now at least 3 characters.

Where appropriate defines have been replaced with constants, enums, or inline functions.