Notes on allocate (and some minor refactoring)
As often with C, the implementation of a simple idea is elegant but requires some thought.
That is, when deriving the idea from the implementation.

freelist is a linked list, and alloc searches and mutates the list.
Mutation, in particular, by splitting nodes into a pair of smaller nodes and inserting the fresh node.

I found this a little easier to follow by using pointer offsets and dereferencing, rather than indexing.
And, by removing the `prev` variable.

It's also tempting to refactor the outer loop, as it's run only once at the moment.
Though, I expect there may be some interest in repeat calls to collect, or applying different strategies.