Broad revisions to listmachine, exercise 10.2
After revisions to the data types used, the list machine failed to run on a handful of examples.
That is, most examples.
And, in each case the machine failed when attempting to execute a CAR instruction.

The list machine was implemented with some pointer mangling.

In particular:

```
the garbage collector can easily distinguish an integer (whose least significant bit is 1) from a reference (whose least significant bit is 0)
```

Is it the case that the least significant bit of a pointer is *always* 0?
Perhaps, though I couldn't find a source.

So, easier to refine away the mangling.

This is done by using a struct in place of raw data which stores raw data along with a type discriminant.
Types used are INT, PTR, and HDR.
The data could be a union, but everything had already been set to use intptr_t already.

Overhead, of course, as the struct is larger and words need to be built when executing the program.
Still, things now work.

Returning to the question above, then, it seems not.

Also included is an initial pass at 10.2.