Exercises 10.3 and 10.4
Mostly straightforward.

Sweeps are done from the start of the heap to the end, so when a blue or white block is found:

The freelist is mutated so the block is first.
A loop is set up which expands the head of the freelist to include the next block if it is blue or white.

That's all, it's quite simple.

Exercise 10.3 is a special case where a break is inserted so at most one instance of the loop happens.

Elsewhere:

inline functions corresponding to removed defines now take word_t pointers, to avoid unnecessary copies given the larger struct size.
In addition, const markers are introduced where possible.

Parameters have been updated to use tag_t where relevant, and tag_t now includes TagFree, allowing a distinction between free blocks and cons blocks.

A fresh assert function has been added to ensure a block type.

Things seems okay, though setting up F# tests is TODO.