Cont setup
Examples from example files recast as tests.

To help with tests a few of the run functions have been updated to take an `out` parameter in order redirect print statements to a string.

Some additional comments on a few on the Icon examples, and an additional example (e4x) to highlight the role of Fail.