Exercise 11.5(ish), split tests
Icon, Contfun, and Contimp tests are split into separate files to avoid module juggling.

A basic fib example for exercise 11.5.
Define fib to raise an error on a negative argument and then define fib on the naturals as usual or on the integers with the condition that any negative index returns 1.