Exercise 12.1, and some refactoring
Exercise:

No notes on the impl.
For tests, filtering to ensure the instruction has been removed seems preferable.
It's possible further optimisations could revert the compilation to an IFZERO, though at least the test only rests on one instruction, and not the full program.

Refactoring:

As it may be useful to make calls to call_machine the code has been split to a separate module.