Exercise 12.4
Quite straightforward.
I should have tried without the hint, as it was very simple to walk through the outline in reverse.
The structure of the code has been refined to match the hint.
Whether to overwrite C is a design choice, but it seems helpful to clarify that the continuation is being built.