Block local declaration assignment continuation
Quite surprised this was not an exercise.

The code is similar to the previous compiler.
Still, the interesting insight is in how variables are handled.
The issue, specifically, being that access to a variable is examind prior to allocation when working backward.
And, the solution implemented is to take two passes at a block.
Both passes generate the same code, though the first pass discards the code and keeps only the environment.
The second pass then uses the environment from the first pass.

The major difficulty was recognising the need to decrement the stack pointer after assignment, as otherwise a the value remains on the top of the stack ready to cause issues.