Exercise 12.4
The relevant optimisation seems to be avoiding accessing the value of a[i] twice.
And, in this case, the optimisation generalises to any acc = acc op expr / expr op acc instance.

The code is fairly simple, amounting to an access then duplication.

And, following the hint the optimisation is applied by inspecting assign expressions rather than the continuation.