MicroC initial commit
In place of X86 the goal is to explore a related task: Compilation of MicroC to LLVM.

An initial task is to find a suitable way to access LLVM bindings.
OCaml seems viable, F# may be viable, and there are various other options.
C++ is chosen, for a variety of reasons, though primarily documentation.

The first task is setting up an AST for MicroC is C++.

This is far less straightforward than with a ML language, though included is something which seems to work ok.

The second task is setting up bison and flex.

This is mostly straightforward, with the most difficult part being the realisation that the docs are very helpful and provide a good foundation.
The grammar Mostly everything else is the same as in F#, so this was a nice tooling related task.
Though, this setup highlighted some shift-reduce conflicts derived from OP_ASSIGN rules which have been fixed.

The not straightforward things are noted in the readme (mostly to do with types).

Handling of for constructs is TODO.

Otherwise, things look in reasonable shape.