Collatz in LLVM
Using fibonacci as a template this was *mostly* straightforward.

Mostly, as understanding blocks and termination took some time, and it's easy to overlook where variables occur when learning a framework.
Still, the IR looks reasonable, and various tests return as expected.

I expect there's a way to directly check booleans, though for now equality is ok.