to_string refactor
Primarily, at least.
The methods have been collected in a single cc file for ease of most everything.

Elsewhere, as_* methods have been removed as they have no use.
Or, unless I'm missing something…
The issue being that as_* methods are intended to be called on a reference to a base class.
However, in this case that as_* attributed to the base is used.
The alternative, then, is to make each instance virtual and specify various failed casts, or case to the desired type some other way.
Some other way seems easier, here.