Additional string formating
Mostly seems ok.

if / then and while are now represented.

A mostly thoughtless approach to indentation has been implemented.
To revised if it turns out to be a mess.
Though, seems ok so far.

Use of fmt has been replaced by std, as the relevant parts of std are derived from fmt, and the only missing item is println.