Unified blocks
Functions now store a pointer to a block for the body of the function.
This allows blocks to be handled in a uniform way.

For parsing, this introduces some issues.
Blocks were parsed as statments, and a handle to a stmt in an fn isn't too helpful.
So, blocks are initially pk'd to a BlockHandle and then abstracted when parsed as a stmt.