Foundations
A large commit, and perhaps a little overdue…

The *foundations* of LLVM codegen have been setup.

Broadly, efforts mostly follow the Kaleidoscope tutorial, though not too closely.

Types have typegen.
Code has codegen.

When things work the main file reads some microC, generates IR and then executes using JIT.

Implemented so far is typegen, and the basics of most codegen with the exceptions of:

- control flow (if/else, while, etc).
- access of index, deref.
- various operators.
- anything details wrt. blocks

codegen uses an LLVMBundle, containing a context, module, builder, etc.

In order to simplify codegen the bundle also contains maps for operators, variables, and other things.

As a slight extension to microC, external `base` functions can be set.
For the moment printf is included in order to implement printi and println.
For these in parituclar, every option seems to have many drawbacks, so...

A design choice was made to implement return blocks for functions.
So, every function is split (at least) into entry and return blocks, where a return breaks to the return block.
If the function has a return value, the value is stored before the break and read at the start of the block.

So, broadly, at the moment simple int assignments, basic ops, and returns work.

e.g. microC such as:

int f(int n) {
  printi n;
  return 4 * n;
}

void main() {
  printi f(2);
}

works out okay, though without much polish.

At present, access instructions are an issue.

The issue is pointers are opaque, but offsets require type information.
And, generally, type information may not be availably locally.

For example:

a[i];

So, some deduction needs to happen.
This, then, seems a reasonable place to commit...