Shift pk functions to driver methods
In order to support LLVM codegen it seems (at least for now) that annotating access nodes in the AST with relevant type information should work ok.

Implementation, then, seems easiest via tracking an env and adding the annotation via the env while parsing.
Handling the env through the driver is simple.
Though, various pk functions will then need access to the driver.

So, all pk funcions are now methods of the driver.
As the methods are no longer namespaced functions, clashes are an issue.
To resolve this each pk now contains information about the struct to be made.

Some initial code for the env is also present, and some driver methods have been rearranged.