Blocks and the environment, while parsing
Handling the environment while parsing is not straightforward.

The simple solution I had in mind was to nest envs whenever a block was formed.
Though, this isn't particularly elegant.
Instead, the solution settled on is to have a simple map from vars to handles, and expand how blocks are handled.
In particular, when adding a var to a block it is simple to determine whether it is shadowing an existing var, and to store the shadowed var to be swapped back when the block ends.
And, it is very easy to determine when a block ends as it is consumed by some else in the AST.

Though, this requires a slight departure from parsing from the book, as it's no longer viable to construct blocks in reverse, as declarations should be handled prior to use.
As this does not introduce any conflicts, it is a simple change.

Elsewhere, which the refined block structure it's a little harder to maintain information about the order of declarations in the source.
So, this is dropped, and instead blocks are represented in a canonical form of:  fresh declarations -> shadowed declarations -> statements.