A bundle of type inference related updates
Roughly, the (partial) implementation is...

Types have been extended with a `deref` method, which dereferences the type.
E.g., deref on an int fails, deref on a pointer goes to the pointed type.
This is partial, for the moment, as pointers aren't handled.

With this, identifying the type accessed through a deref is mostly straightforward.
Get the type of the expression and deref...

With this, an `expr_type` method can be defined for expressions.
This, mostly, returns the type of the expression when evaluated, and so may include dereferencing.
At the moment it's partial and no checks are in place to ensure types are ok.

Still, type information for GEPs is closer.
As, with these methods it's now mostly possible to the type of an expression, and hence the type of register required for the load.

And, as part of the above fns are now stored in the env.
This was a overlooked earlier.

To accomodate this the env is generalised to hold DecHandles, as fns can be shadowed (as a choice).

Also included is simplification of shadowed storage to lookup the var.
Minor, and optimisation is of no concern at the moment.