Initial pass at global codegen
It took some time to track down a hint that globals must be initialised.
Quite some time.
And, for the moment only ints are initialised.
And, the implementation is quite messy, as defaults really should have a unified source, perhaps as a method.
In any case, the generated code is roughly as expected.

In addition, named values are abstracted to store Values, and is no longer cleared when enterting codegen for a function (as globals may be present).