Global variables, an additional step
The implementation of the env while parsing was a little confused.
Specifically, by updating the env in multiple places.
This has been fixed.

For codegen, the idea is to replicate parsing, roughly.
When examining a block shadowed vars are temporarily stored, and then restored on exit.
Likewise, fresh variables will be cleared.

Quickly, shadowed vars work as expected, on small tests via JIT...