Conditionals, first pass
The structure follows Kaleidoscope for the most part.
The only significant diversion is omission of some details regarding returning a value, which aren't required for microC, but which may be introduced later.

The comparison is also a little more explicit.

In addition, the handling of return blocks has been revised.
Here, given the realisation that there's no need to insert the return block when created, and that things are generally much easier if insertion of the block is delayed (as then there's no need to traverse iterators, etc. in order to find the last non-return block).

Small tests, as with everything else so far...