AST: Conditionals branches to blocks
The basic handling of conditional codegen does not extend well to early returns.
To handle this, and potentially other things, additional information about the branches is useful.

For example, when building the AST it's possible to include information about flow that codegen can use.

To simplify things a little, this information will be attributed to a block.
And, so, the branches of a conditional will always be blocks.

Internally, at least.

In other words, whenever a non-block branch is parsed, the statement is cast to a block before inclusion in the AST.

Also fixed is an outdated complete_data call (`_unsafe`) added.