Improved conditional handling
Following from the previous commit, some metadata on blocks is tracked.
Specifically, counts on (early) reutrns and fall throughs.
While these aren't strictly required (I don't think) for the current approach, they are useful to avoid creating redundant return statements.

Speaking of the current approach...

A conditional `end` block is always created, which handles everything following the conditional.
Though, an `else` block is created only when required.

This seems to perform as expected, though I expect it fails with void fns.

A todo (perhaps, and so noted here) is avoiding creation of the `end` block when all branches make early returns.
For the moment the issue is that code for the fn return is always generated...