Return improvements
Return blocks are now created only when there is *some* pass through to the end of the fn.

And, with this, it's easy to determine whether to return or store and break by inspecting whether a return block / allocation has been set.

Though, that's only part of it...

Whether or not a block returns is important, but in the case of fns it's also important to know whether the return is `final`.
Or, as for the moment used. `in scope`.
As, if the return is final then (conservatively) setup for the return block should be performed, and otherwise the fn returns on all paths.

`In scope` works roughly the same as more strict check on whether the return is final *and* in the current scope.

The check for `in scope` is simple, as a return ends the current block (even if more is parsed), and it's only set when the non-final return flag has not been set.

On reflection, it might simplify many things to simply no parse various things to the AST (e.g. when final is set on a block), though that's for some other time...

The fn codegen sets up a return block whenever there's a final return to be made, and similarly if / then block generation depends on whether there's a branch, etc.

This still leads to some redundancy if there's a single use of the final return, though is an improvement on what was generated previously.

Counts of reuturns / pass throughs could be removed, but they may be useful to keep...?