While codegen, first pass
Somewhat similar to conditionals, though simplified to the extent that there's no `else` block and so it's never the case that a while `returns` in general (as the body may never be entered).

Though, so caution is required as the body of a while loop may return, and as such there may be a conflict with the return instruction and the break (/ loop) instruction.

Elsewhere, return / pass counts have been updated to methods.
And, a fix for incorrectly marking while statements as if statements has been fixed.

A "Fmt" header has been added, with source, to hold formatting of enums, etc.