Fn revisions
The bundle has been reorganised a little.

The source now contains a `codegen` folder, with the llvm bundle, and related functionality.
The codegen methods are moved to an `AST.cpp` file, and the handling of op generation has been revised.

For op generation, each type of op now has a separate file, and the maps are extended with helper functions.
In turn, each fn is set by calling a function which takes a bundle and returns a builder of the appropriate type.
This isn't particularly efficient, nor elegant, though it seems sufficiently simple.