Some progress on address and arrays
Array declaration does not, on reflection, ensure default allocation.
It's possible, but I don't think it's particularly interesting, nor in the spirit of microC.
Perhaps default initialisation in general isn't in the spirit, though as it's required for global constants due to LLVM constraints it makes sense to extend to local declarations also.

So, declaration has been revised to branch on whether or not an array is being declared.

Indexing works fine (small tests), though pointer arithmetic does not.
Similarly, taking the address of a pointer works fine with dereferencing, but that's all so far.

Some progress, then...

In addition, exceptions replace panics in codegen.