Codegen for ops revised to take AST nodes
The motivation is codegen which requires type resolution.
E.g., addition between ints and pointer arithmetic.

At some point codegen diverges, and it makes (at least some) sense to diverge at the point of ambiguity --- the operator.
And, node handles contain useful information for resolving ambiguity.
In particular, for the case of pointer arithmetic, a way to identify offset requirements.