Foundations for allocation revision
In rough outline the idea is to revise the AST with nodes for allocations, and for an expression to hold an allocation.
These nodes then allow for simple de/referencing, as the action taken on a allocation will not be bundled with the allocation.
In particular, the access can be removed, dereferencing can be handled as a unary op, etc.
Complexity is introduced by imposing on other expressions the task of handling the allocation (typically loading).

Still, overall things should be a little simpler.
For example, indexing can be revised to an expression which generates a pointer, and nothing else is required.

With respect to implementation, the rough idea is to use the existing Dec class as the allocation class, and introduce a declaration statement which contains an allocation.
The expression class will be extended, and the parser will be revised to generate appropriate nodes.

let's see...