Checkpoint, of sorts
A few days pass under the weather and things aren't so clear in memory.
In any case...

Parsing has been revised to handle *most* access instances in a generic way.
That is, `&` and `*` have been introduced as unary ops, and are parsed to these ops.

The idea is that any allocation is kept as a pointer and only loaded 'when required'.
And, for the moment, at least, 'when required' is determined by examing the type of expression and type of llvm value.
E.g. if the value is a pointer and the expression some data type then a load should happen.
To do this a `ensure_loaded` method has been added to the LLVMBundle class.
Strictly, there's no need for the class at the moment, but there's no forseeable instance when independence from the class would be useful.

Elsewhere access nodes no longer carry mode information.

And, various instances of printf debugging are present...