Resolve parse conflicts
In a previous commit `-` was introduced as a general unary operator.
This led to conflicts with parsing constants, -(const) vs (-const).
This is resolved by removing the rule for constants.
And, for clarity, the `Const` rule has been renamed to `AtomicConst`.

In broader context, if not sufficiently clear already, this is another example of using the AST to encode useful information that doesn't (easily) generalise to expanded parsing, similar to access nodes.