Foundations for type resolution
Information about types is required for codegen, esp. for variables.
Prior to this commit relevant work was done during parsing by maintaining an env.
(This code is still present, for the moment.)
Though, the work was partial, as the env did not track information about parameters in an fn.

All this could be done at parsing, though for simplicity it seems preferable to do most of this work as a second `type resolution` pass.
And, the goal is for this pass to generate type information for all expressions, not only variables.

To simplify type resolution a fresh `type_resolution` method has been added to the Node class, with stub implementations for each node.

And, while doing this node classes have been cleaned up a little...