Checkpoint
The headline is that params are now tracked in the env, though some cleanup is required.

Though, plans change again.

Type resolution was planned as a second pass.
Though, on revision it will remain part of parsing.
The motivation for this is blocks.
At present blocks track whether variables are shadowed or fresh.
To do this, blocks take-from and restore-to the env in order to avoid a DAG of envs.
On reflection, a DAG may have been wiser, but that's an aside.
The key point is that when tracking declarations it's very little additional work to track types.

Further, if types are set during parsing then type information can be required for the construction of a node, avoiding the possibility of nullptr initialisations.

So, many things still to do, but the path seems a little clearer.

In terms of significant changes:

- Parsing of fns is separated into prototypes and declarations.
  This allows adding all the params to the env, and will also allow for recursive calls (at some point).
- ParamVecs store params as token / type pairs, rather than type / token for some consistency.
- Fns need to be finalised to remove params from the env, similar to blocks.
- A return_type method has been added to fns for clarity.

Various things are incomplete / broken.