Checkpoint, ex1.c from the book JIT (almost) ok
The only thing missing is the println, though that can wait.

In short, the execution engine has been revised a little to handle arguments.
In turn, it no longer uses the interpreter.
Without the interpreter the calls to printf failed, and so those have been replaced with calls to a custom printi function (which wraps printf).

In long, I'm not even sure where to begin with understanding the issue, and it may be preferable to use a different HJIT, if kalaediscope is to be followed.
Regardless, this works, and it's been a headache...