AST documentation, protected expr types
Small tidying.

Various parts of AST.hpp have been documented, with some part reordered for clarity.

In addition, the `typ` member of expressions is protected, with some access updated to method calls.
The method calls continues to check for the presence of a type, though there should no longer be any need for this.