Type simplification
Int, Char, and Void types are no longer variants of Data type.

The motivation is simplification of codegen, etc.
As, given microC has few types there's not much to be gained by imposing a hierarchy.

A significant consequence is type completion, as it's no longer possible to mutate a fixed data type node.
So, a temporary fix is in place.

In the future it's possible pointer and index types will be merged, as the distinction is weak.