Merge pointer and index types
To further simplify things and reduce repetition.

The key observation is that during declaration there's no ambiguity due to `*` vs `[]`.
And, at any other time ambiguity is resolve by whether or not the type has an associated size (though this should really be length or ...).

Well, mostly, at least.
This leaves one case: parameter VLAs.
In C these are pointers, of course, so there's no issue wrt. functionality.
Though, it is some loss.
One option is to give these size zero, internally.

A todo is ensuring loads of pointers, as here pointers should be used in all cases.
Though, this issue was present prior to the merge.