Pointer type tidying
Index renamed pointer, fields made private, and some renaming.

Nothing too important.

As a continuation of the last commit...

One other issue is the mapping between AST types and LLVM.
As, before the mapping was one to one, and is now one to many (two).
Though, I don't think this is too much of an issue, as generally (esp. with other languages) I doubt it's viable to maintain.