Checkpoint
Some progress.

ensure_access is now access, and has changed a little.

With respect to codegen, there's no immediate action that can be taken for * and & ops.
Consider:

*(b + 1) = *(b + 2)

In words, the memory at b + 1 should contain the value of the memory at b + 2.
For IR, the lhs should be a pointer and the rhs should be a value.

So, what settles how to act is the equality operator.
Keep the address of the lhs and load the value of the rhs.

Previously, ensure_load attempted to do this, but in a messy way.
The idea was to check to see if the IR was a pointer, and whether a pointer was in the AST.
A hack, but one that worked for simple types.

The more general approach is to inspect the expression to see whether it's something that dereferences.
Though, this is (in part) only viable now as each expression is annotated with a type.

For the moment only addition has been given attention, though other things seem to work (print, etc.).