Simplify access (again)
With the changes made to access there's no need, in general, to attempt access on a value.
Instead, access can be directly called on an expression.
That is, at every point in the source where access was used, it was immediately preceeded by codegen for the value.

So, an overload has been added which does this and each use has been simplified.
In future the fresh overload may be replace the original method.

Oh, and 'access' is used over 'access_if' given the role this method now has.