Revised primitive fns
In place of codegen, primitive fns now inherit from a FnPrimative struct.
This provides a single location for name, codegen, etc.

Of note, the implementation of the functions has been moved to accompany the struct, and the struct now returns a LLVM ok pointer.
Though, it seems this pointer is not needed.
I don't, at this moment, have much desire to figure out the specifics of external function calls, so things have only been rearranged and commented.