Types, envs, and fn issues, checkpoint
The idea of tracking an evaluation type for each expression wasn't so good.
The issue, here, again, is that the type of interest depends on the context of evaluation.
Specifically, with respect to equality.

So, broadly, the goal is to track the type of the expression, without any implicit dereferencing, etc.

This is partially complete.

On the way to implementing the above a further diversion is required.
This time, due to MicroC from the book.
As, in the book there is no distinction between ints and pointers.
Or, slightly more accurately, both ints and pointers have the same internal representation.

So, functions such as `print` can be called on both.
This is not the case in the current ast / llvm representation and I'd like that to remain.
As such, some conversion is required.

In turn, this requires --- I think --- additional information when calling functions.
Specifically, information about arguments to allow inference of whether some coercion is required.

This information isn't too hard to track, but requires some revisions.

This commit, in particular, focuses on initial refinments to envs in order to track additional information.

The idea is to expand fns in the env to hold information about argument and return types, and to use this for inference.

To do this, the AST env needs to be available during codegen, so has been moved to the LLVMBundle struct.
The struct is part of the driver, so there's no particular issue here.
Though, it would be cleaner to transform a driver into an LLVMBundle, given how complex things are trending.
Regardless, the key insight is that with the way the env is handled, at the end of parsing it will contain all functions and global variables, as variables in blocks are removed when exiting a block.
Useful.

Then, it's a matter of expanding the information about functions.
And, in anticipation, fns and vars have been split in the env.

Well, some progress...